hydra-proj/pyproject.toml
# build section


# project metadata section
[project]
name = "hydra-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "pandas>=2.3.2",
]


# tool section
[tool.setuptools.packages.find]
where = ["src"]
